Avoid warning in jsapi-tests/tests.h due to printf type mismatch. No bug, rs=Waldo.
--HG--
extra : rebase_source : 22429e5178fa794fc06da843ba8046a2678ca114

+++ js/src/jsapi-tests/tests.h
@@ -202,7 +202,7 @@ public:
             if (s)
                 msg += JS_GetStringBytes(s);
         }
-        fprintf(stderr, "%s:%d:%.*s\n", filename, lineno, msg.length(), msg.begin());
+        fprintf(stderr, "%s:%d:%.*s\n", filename, lineno, (int) msg.length(), msg.begin());
         msgs += msg;
         return false;
     }
